Mooncake (25)
原来是类型出错了，奶奶的。 还有谁他妈知道这是谈心算法？

+++ 1070.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 struct Node
 {
-	int inventory;
+	double inventory;
 	double price;
 	double one_price;
 };
